third_party/blink/web_tests/external/wpt/css/css-values/sin-cos-tan-serialize-expected.txt

This is a testharness.js-based test.
FAIL 'cos(0)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'cos(0)' should be valid in opacity. got disallowed value ""
FAIL 'scale(cos(0))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(cos(0))' should be valid in transform. got disallowed value ""
PASS 'cos(0)' as a computed value should serialize as '1'.
FAIL 'scale(cos(0))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(cos(0))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'sin(0)' as a specified value should serialize as 'calc(0)'. assert_not_equals: 'sin(0)' should be valid in opacity. got disallowed value ""
FAIL 'scale(sin(0))' as a specified value should serialize as 'scale(calc(0))'. assert_not_equals: 'scale(sin(0))' should be valid in transform. got disallowed value ""
FAIL 'sin(0)' as a computed value should serialize as '0'. assert_equals: 'sin(0)' and '0' should serialize the same in computed values. expected "0" but got "1"
FAIL 'scale(sin(0))' as a computed value should serialize as 'matrix(0, 0, 0, 0, 0, 0)'. assert_equals: 'scale(sin(0))' and 'matrix(0, 0, 0, 0, 0, 0)' should serialize the same in computed values. expected "matrix(0, 0, 0, 0, 0, 0)" but got "none"
FAIL 'tan(0)' as a specified value should serialize as 'calc(0)'. assert_not_equals: 'tan(0)' should be valid in opacity. got disallowed value ""
FAIL 'scale(tan(0))' as a specified value should serialize as 'scale(calc(0))'. assert_not_equals: 'scale(tan(0))' should be valid in transform. got disallowed value ""
FAIL 'tan(0)' as a computed value should serialize as '0'. assert_equals: 'tan(0)' and '0' should serialize the same in computed values. expected "0" but got "1"
FAIL 'scale(tan(0))' as a computed value should serialize as 'matrix(0, 0, 0, 0, 0, 0)'. assert_equals: 'scale(tan(0))' and 'matrix(0, 0, 0, 0, 0, 0)' should serialize the same in computed values. expected "matrix(0, 0, 0, 0, 0, 0)" but got "none"
FAIL 'calc(sin(0) + cos(0) + tan(0))' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'calc(sin(0) + cos(0) + tan(0))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(sin(0) + cos(0) + tan(0)))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(calc(sin(0) + cos(0) + tan(0)))' should be valid in transform. got disallowed value ""
PASS 'calc(sin(0) + cos(0) + tan(0))' as a computed value should serialize as '1'.
FAIL 'scale(calc(sin(0) + cos(0) + tan(0)))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(calc(sin(0) + cos(0) + tan(0)))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'calc(sin(0) + 0.5)' as a specified value should serialize as 'calc(0.5)'. assert_not_equals: 'calc(sin(0) + 0.5)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(sin(0) + 0.5))' as a specified value should serialize as 'scale(calc(0.5))'. assert_not_equals: 'scale(calc(sin(0) + 0.5))' should be valid in transform. got disallowed value ""
FAIL 'calc(sin(0) + 0.5)' as a computed value should serialize as '0.5'. assert_equals: 'calc(sin(0) + 0.5)' and '0.5' should serialize the same in computed values. expected "0.5" but got "1"
FAIL 'scale(calc(sin(0) + 0.5))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'. assert_equals: 'scale(calc(sin(0) + 0.5))' and 'matrix(0.5, 0, 0, 0.5, 0, 0)' should serialize the same in computed values. expected "matrix(0.5, 0, 0, 0.5, 0, 0)" but got "none"
FAIL 'calc(cos(0) + 0.5)' as a specified value should serialize as 'calc(1.5)'. assert_not_equals: 'calc(cos(0) + 0.5)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(cos(0) + 0.5))' as a specified value should serialize as 'scale(calc(1.5))'. assert_not_equals: 'scale(calc(cos(0) + 0.5))' should be valid in transform. got disallowed value ""
FAIL 'calc(cos(0) + 0.5)' as a computed value should serialize as '1.5'. assert_equals: '1.5' should round-trip exactly in computed values. expected "1.5" but got "1"
FAIL 'scale(calc(cos(0) + 0.5))' as a computed value should serialize as 'matrix(1.5, 0, 0, 1.5, 0, 0)'. assert_equals: 'scale(calc(cos(0) + 0.5))' and 'matrix(1.5, 0, 0, 1.5, 0, 0)' should serialize the same in computed values. expected "matrix(1.5, 0, 0, 1.5, 0, 0)" but got "none"
FAIL 'calc(tan(0) + 0.5)' as a specified value should serialize as 'calc(0.5)'. assert_not_equals: 'calc(tan(0) + 0.5)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(tan(0) + 0.5))' as a specified value should serialize as 'scale(calc(0.5))'. assert_not_equals: 'scale(calc(tan(0) + 0.5))' should be valid in transform. got disallowed value ""
FAIL 'calc(tan(0) + 0.5)' as a computed value should serialize as '0.5'. assert_equals: 'calc(tan(0) + 0.5)' and '0.5' should serialize the same in computed values. expected "0.5" but got "1"
FAIL 'scale(calc(tan(0) + 0.5))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'. assert_equals: 'scale(calc(tan(0) + 0.5))' and 'matrix(0.5, 0, 0, 0.5, 0, 0)' should serialize the same in computed values. expected "matrix(0.5, 0, 0, 0.5, 0, 0)" but got "none"
FAIL 'calc(sin(infinity))' as a specified value should serialize as 'calc(NaN)'. assert_not_equals: 'calc(sin(infinity))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(sin(infinity)))' as a specified value should serialize as 'scale(calc(NaN))'. assert_not_equals: 'scale(calc(sin(infinity)))' should be valid in transform. got disallowed value ""
FAIL 'calc(sin(infinity))' as a computed value should serialize as 'NaN'. assert_equals: 'NaN' should round-trip exactly in computed values. expected "NaN" but got "1"
FAIL 'scale(calc(sin(infinity)))' as a computed value should serialize as 'matrix(NaN, 0, 0, NaN, 0, 0)'. assert_equals: 'matrix(NaN, 0, 0, NaN, 0, 0)' should round-trip exactly in computed values. expected "matrix(NaN, 0, 0, NaN, 0, 0)" but got "none"
FAIL 'calc(cos(infinity))' as a specified value should serialize as 'calc(NaN)'. assert_not_equals: 'calc(cos(infinity))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(cos(infinity)))' as a specified value should serialize as 'scale(calc(NaN))'. assert_not_equals: 'scale(calc(cos(infinity)))' should be valid in transform. got disallowed value ""
FAIL 'calc(cos(infinity))' as a computed value should serialize as 'NaN'. assert_equals: 'NaN' should round-trip exactly in computed values. expected "NaN" but got "1"
FAIL 'scale(calc(cos(infinity)))' as a computed value should serialize as 'matrix(NaN, 0, 0, NaN, 0, 0)'. assert_equals: 'matrix(NaN, 0, 0, NaN, 0, 0)' should round-trip exactly in computed values. expected "matrix(NaN, 0, 0, NaN, 0, 0)" but got "none"
FAIL 'calc(tan(infinity))' as a specified value should serialize as 'calc(NaN)'. assert_not_equals: 'calc(tan(infinity))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(tan(infinity)))' as a specified value should serialize as 'scale(calc(NaN))'. assert_not_equals: 'scale(calc(tan(infinity)))' should be valid in transform. got disallowed value ""
FAIL 'calc(tan(infinity))' as a computed value should serialize as 'NaN'. assert_equals: 'NaN' should round-trip exactly in computed values. expected "NaN" but got "1"
FAIL 'scale(calc(tan(infinity)))' as a computed value should serialize as 'matrix(NaN, 0, 0, NaN, 0, 0)'. assert_equals: 'matrix(NaN, 0, 0, NaN, 0, 0)' should round-trip exactly in computed values. expected "matrix(NaN, 0, 0, NaN, 0, 0)" but got "none"
Harness: the test ran to completion.
